+++ ceammc/ext/src/noise/chaos_gbman.cpp
@@ -16,23 +16,15 @@
 
 #include <cmath>
 
-//constexpr double PI = 3.141592653589793;
-//constexpr double TWO_PI = 2 * PI;
-//constexpr double RECIP_PI = 1 / PI;
-//constexpr double RECIP_2PI = 1 / TWO_PI;
-
 constexpr t_float DEF_X = 1.2;
 constexpr t_float DEF_Y = 2.1;
-
-//static inline t_sample wrap_pi(t_sample x)
-//{
-//    return (x - PI) * RECIP_2PI;
-//}
+constexpr t_float MIN_FREQ = 0.001;
+constexpr long MIN_PERIOD = 1;
 
 static inline double gbman_next(double& xn, double& yn)
 {
     //    x(n+1) = 1 - y(n) + |x(n)|
-    //            y(n+1) = x(n)
+    //    y(n+1) = x(n)
     double xn0 = xn;
     xn = 1 - yn + std::fabs(xn0);
     yn = xn0;
@@ -52,6 +44,13 @@ void ChaosGbman::onBang()
     floatTo(0, gbman_next(xn_, yn_));
 }
 
+void ChaosGbman::dump() const
+{
+    OBJ_DBG << "x: " << xn_;
+    OBJ_DBG << "y: " << yn_;
+    BaseObject::dump();
+}
+
 ChaosGbmanTilde::ChaosGbmanTilde(const PdArgs& args)
     : SoundExternal(args)
     , xn_(positionalFloatArgumentT(0, DEF_X))
@@ -63,10 +62,11 @@ ChaosGbmanTilde::ChaosGbmanTilde(const PdArgs& args)
 
 void ChaosGbmanTilde::processBlock(const t_sample** in, t_sample** out)
 {
-    t_float freq = in[0][0];
+    const t_float freq = in[0][0];
+    const size_t SR = samplerate();
     t_float output = xn_;
-    t_float call_period = (freq < samplerate()) ? (samplerate() / std::max<t_float>(freq, 0.001)) : 1;
-    call_period = std::max<t_float>(call_period, 1);
+    long call_period = (freq < SR) ? (SR / std::max<t_float>(freq, MIN_FREQ)) : MIN_PERIOD;
+    call_period = std::max<long>(call_period, MIN_PERIOD);
 
     const auto BS = blockSize();
     for (size_t i = 0; i < BS; i++) {
@@ -80,8 +80,25 @@ void ChaosGbmanTilde::processBlock(const t_sample** in, t_sample** out)
     }
 }
 
+void ChaosGbmanTilde::dump() const
+{
+    OBJ_DBG << "x: " << xn_;
+    OBJ_DBG << "y: " << yn_;
+    SoundExternal::dump();
+}
+
 void setup_chaos_gbman()
 {
     ObjectFactory<ChaosGbman> obj("chaos.gbman0");
-    SoundExternalFactory<ChaosGbmanTilde> obj_tilde("chaos.gbman0~", OBJECT_FACTORY_DEFAULT | OBJECT_FACTORY_MAIN_SIGNAL_INLET);
+//    obj.setDescription("Gingerbreadman map chaotic generator");
+//    obj.setCategory("chaos");
+//    obj.addAuthor("Serge Poltavski");
+//    obj.since("0.9");
+
+    SoundExternalFactory<ChaosGbmanTilde> obj_tilde("chaos.gbman0~",
+        OBJECT_FACTORY_DEFAULT | OBJECT_FACTORY_MAIN_SIGNAL_INLET);
+//    obj_tilde.setDescription("Gingerbreadman map chaotic generator");
+//    obj_tilde.setCategory("chaos");
+//    obj_tilde.addAuthor("Serge Poltavski");
+//    obj_tilde.since("0.9");
 }